meta-ads-mcp-server/requirements.txt
# Meta Ads MCP Server Dependencies

# MCP Framework
fastmcp>=2.6.1

# Meta/Facebook API
facebook-business>=20.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP Requests
requests>=2.31.0
httpx>=0.24.0

# Environment Management
python-dotenv>=1.0.0

# Validation
pydantic>=2.0.0

# Date/Time
python-dateutil>=2.8.0

# Caching
redis>=4.5.0

# Logging
structlog>=23.0.0

# JSON Processing
orjson>=3.8.0

# Image Processing (for ad images)
Pillow>=10.0.0

# Rate Limiting
ratelimit>=2.2.0

# Async Support
aiofiles>=23.0.0
asyncio-mqtt>=0.16.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools (optional)
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0 